This is another update to creating a interface abstraction layer to the ICSharpCode.SharpDevelop solution.  The majority of this update is just creating a series of interfaces for each class, thus eventually enabling the classes to be substituted in and out in a gradual manner instead of having to perform core refactorings.
Most of the work went without a hitch.  ItemType operators are not being used any longer; they do not play nice with the C# spec, so where the reference was being checked with it when determining project dependencies, the == equals operator was replaced with a call to the Equals method.

There was also one class hierarchy that did not make sense.  In the Master branch, SolutionFolder inherits AbstractSolutionFolder and ISolutinoFolderContaienr.  Expanded through inheritence, the list is LocalizedObject, ISolutionFolder, ISolutionFolderContainer, AbstractSolutionFolder.  The problem with this logic is there naming consistency issue between the classes and interfaces.  An ISolutionFolder is an subset of AbstractSolutionFolder but the implementation of SolutionFolder is superset of AbstractSolutionFolder.  Therefore, when you switch to an interface implementation, where interfaces map to the classes, a variable that was declared as SolutionFolder, moves down the type hierarchy when the type is switched to ISolutionFolder.  The result is the hierarchy breaks.

The solution to this is that SolutionFolder should have been named SolutionFolderContainer.  This class does not exist in the type hierarchy of the solution till now.  After adding this type, and switching references from SolutionFolder to SolutionFolderContainer, the problem was solved, and the hierarchy is again consistent.

Last, the last commit did not commit the new interfaces.  This one has them all.  Enjoy!